+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### vNext
+
+- Speech recognition now uses the correct sample rate in Firefox.
+
 ### v4.0.2
 
 - Stopping the microphone now stops the underlying `MediaStreamTrack`, removing the red 'listening' icon.